WiringAssistant/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
numpy==2.1.2
opencv-python==4.11.0.86
Pillow==10.4.0
